Redisearch indexing (#558)
* added redisearch as a dependency

* updated makefiles

* update redismodule api include of redisearch api header

* force add redisearch_cflags.cmake, required by redisearch cmake

* added -ldl to LDFLAGS required by dlopen

* fulltext search

* Fix RediSearch iterator allocation

* Specify library links after dependencies in unit test compilation

* Use updated RediSearch version

* Update RediSearch library name in Makefiles

* Switching exact match index to RediSearch WIP

* drop full-text index

* refresh index schema, utilize index scan

* index crud ops

* persistency WIP

* updated redisearch version

* WIP

* free safe operations, index scan reset iterator

* test suite pass

* rebased

* rebase fixes

* WIP

* added redisearch as a dependency

* updated makefiles

* update redismodule api include of redisearch api header

* force add redisearch_cflags.cmake, required by redisearch cmake

* added -ldl to LDFLAGS required by dlopen

* fulltext search

* Fix RediSearch iterator allocation

* Specify library links after dependencies in unit test compilation

* Use updated RediSearch version

* Update RediSearch library name in Makefiles

* Switching exact match index to RediSearch WIP

* drop full-text index

* refresh index schema, utilize index scan

* index crud ops

* persistency WIP

* updated redisearch version

* WIP

* free safe operations, index scan reset iterator

* test suite pass

* rebased

* rebase fixes

* WIP

* Create builds edge schemas and properties

* Construct exact indexes

* Commit deletions before printing results

* Don't track new relationship types under labels_added

* Disable CALL...WHERE tests

* missing tag node

* resolve procedure output name

* Use latest version of libcypher-parser

* Re-enable CALL...WHERE tests, update AST validations

* Rebuild CircleCI cache on grammar changes

* Ensure string safety in prev deserialization routine

* Make all procedure tests deterministic

* Remove triemap from module

* Replace all triemaps with rax

* Various fixes

* Disallow building value join ops on non-scalars

* Correct bug in schema serialization

* Fix memory leak

* Don't crash on querying non-existent full-text index

* Documentation updates

* Address PR comments

* updated redisearch version

* PR updates contd

* Tentative solution for schema/index serialization

* Reduce AR_ExpNodes to scalars early if possible

* Serialize index data as part of Schema routines

* PR updates

* Improve SIValue memory ownership interface

* Improved operation responsibility over SIValues

* Reduce unnecessary Projection persistence

* Record_Clone and Record_Merge don't violate SIValue management

* Fix memory leak in complex AR_Exp_Evaluate trees

* Fix memory leak in unary operator conversion

* Only attempt to reduce finalized AR_ExpNodes

* Change AR_EXP_ReduceToScalar return type to boolean

* Better interfaces in updated Record logic

* Use SIValue_MakeVolatile interface to update memory ownership of shallow copies

+++ src/execution_plan/ops/op_delete.c
@@ -16,6 +16,9 @@ void _DeleteEntities(OpDelete *op) {
 	uint node_count = array_len(op->deleted_nodes);
 	uint edge_count = array_len(op->deleted_edges);
 
+	/* Nothing to delete, quickly return. */
+	if((node_count + edge_count) == 0) return;
+
 	/* Lock everything. */
 	Graph_AcquireWriteLock(g);
 
@@ -104,10 +107,25 @@ OpResult OpDeleteReset(OpBase *ctx) {
 void OpDeleteFree(OpBase *ctx) {
 	OpDelete *op = (OpDelete *)ctx;
 
-	_DeleteEntities(op);
+	if(op->deleted_nodes || op->deleted_edges) _DeleteEntities(op);
 
-	array_free(op->nodes_to_delete);
-	array_free(op->edges_to_delete);
-	array_free(op->deleted_nodes);
-	array_free(op->deleted_edges);
+	if(op->nodes_to_delete) {
+		array_free(op->nodes_to_delete);
+		op->nodes_to_delete = NULL;
+	}
+
+	if(op->edges_to_delete) {
+		array_free(op->edges_to_delete);
+		op->edges_to_delete = NULL;
+	}
+
+	if(op->deleted_nodes) {
+		array_free(op->deleted_nodes);
+		op->deleted_nodes = NULL;
+	}
+
+	if(op->deleted_edges) {
+		array_free(op->deleted_edges);
+		op->deleted_edges = NULL;
+	}
 }
